requirements.txt
flask==3.0.0
ollama==0.2.1
requests==2.31.0
python-dotenv==1.0.0